+++ CHANGELOG.md
@@ -11,6 +11,10 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 
 - Observation matrix dashboard: several filter options don't work [#4640]
 
+### Changed
+
+- Add the POST method for Image matrix to handle many OTUs
+
 [#4640]: https://github.com/SpeciesFileGroup/taxonworks/issues/4640
 
 ## [0.56.1] - 2025-12-12